UIIN-3328 move validateAdditionalCallNumbers to util.js
